requirements.txt
langchain==0.3.26
langchain-core==0.3.26
langchain-community==0.3.26

langchain-pinecone==0.2.8
langchain-google-genai==0.1.4

flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0

pinecone-client==3.2.2
google-generativeai==0.7.2

-e .